Multiple panel content showing
When linking into a panel from another set of tabs via an anchor it was possible to see the default active panel as well as the panel your link was in. Hiding after removing the active class takes care of this problem. It also fixes the back button navigating to the default panel instead of back to the tab group you originally came from. Should fix #93