Mobile speedup, no active tab class propagation
Mobile platforms don't click, they use touch events, and through some complicated mess of jQuery, no doubt, they're translated to clicks. The problem is that that takes a very long time.

I bound the tab click handler to 'touchend' as well, and if it ever receives a touchend event it unbinds the click event so that the function isn't run twice.

I'm almost certain there's a better way to do this.

Additionally, I commented out the portion that propagated the active tab class to the li's children, on line 466, 467, and the end of 592.

It's true that the li>a selector doesn't perform as well, but I don't think that the small speedup is worth it compared to using yet another class, especially considering that the selector doesn't need to traverse a complex structure to find li>a, that's the only element li holds usually. 

If you do want that, uncomment 466, 467, end of 592.